Use separate queue for hooks
The previous behavior incorrectly placed hook execution in the same queue
with the database updates. This fixes that issue by correcting that
inconsistency.

The primary reason for wanting separate queues for hooks vs. sequel
messages is that hooks should ideally run sequentially to minimize the
potential for avoid race conditions. The sequel queue allows concurrent
requests, whereas the hooks queue does not.

For https://tickets.puppetlabs.com/browse/RAZOR-642
