+++ spec/app/update_hook_configuration_spec.rb
@@ -5,7 +5,9 @@ require_relative '../../app'
 describe Razor::Command::UpdateHookConfiguration do
   include Razor::Test::Commands
 
-  use_hook_fixtures
+  before :each do
+    use_hook_fixtures
+  end
 
   let(:app) { Razor::App }
